UBO Read Bandwidth profiler (#4270)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/4270

This diff introduces a profiler that obtains the maximum and minimum bandwidth for reading Uniform Buffer Objects (UBOs) from memory, using the following shader, where A is a UBO and B is a writeonly buffer.

  void main() {
    vec4 sum = vec4(0);
    const uint workgroup_width = local_group_size * niter * ${NUNROLL};
    uint offset = (gl_WorkGroupID[0] * workgroup_width  + gl_LocalInvocationID[0]) & addr_mask;

    int i = 0;
    for (; i < niter; ++i)
    {
        sum *= A[offset];
        offset = (offset + local_group_size) & addr_mask;
        ...
        ...
        sum *= A[offset];
        offset = (offset + local_group_size) & addr_mask;
    }

    vec4 zero = vec4(i>>31);

    B[gl_LocalInvocationID[0]] = sum + zero;
}

Through the WorkGroup and LocalInvocation IDs, we ensure that each run accesses its own block of memory. Then, we repeatedly access several unique address locations, using an address mask to wrap around if necessary.
Finally, we make sure to use the `sum` and `i	` variables so that the compiler's optimizer does not flatten the loops.

For a Samsung S22, the bandwidth behaves like this. We can see a decline proportional to the size of the buffer, until it plateaus at 32KB.

{F1759435529}

Comparing it with the Readonly profiler, we can immediately see the superiority in reading speed for UBOs, whenever the hardware is available

Samsung S22
{F1759435825}

Redmi Note
{F1759445004}

Differential Revision: D59776899